Add postinstall configuration
In the postinstall configuration, we add a user to the environment,
assign the proper role to that user for the sa-telemetry
namespace/project, and modify the broker-config for the
openshift-ansible-service-broker so that it will be properly
configured for deployment of ElasticSearch.

Once the broker-config is modified, then the ASB is restarted (rollout
latest). Changes to the nodes are also executed to up the max file
descriptors needed for ElasticSearch deployments.