Deploy Prometheus in 2 deployments
Perform 2 separate deployments of Prometheus so that we can better
control which consumers we're connecting to. The architecture dictates
that we have a blue setup, where 1 prometheus connects to 1 consumer, and
a green deployment that is the same.

Prior to this change we deployed a single Prometheus that connected to
both consumers, but that is not what we want to happen.